Move auth controllers to controllers directory under db file path
